Final code cleanup before the release (see #4998)
Description
-----------

-

Commits
-------

00e9f9c2 Final code cleanup before the release

+++ core-bundle/contao/classes/Theme.php
@@ -730,7 +730,7 @@ class Theme extends Backend
 	private function importExampleWebsite(string $exampleWebsite, bool $preserveData, bool $insertOnly): void
 	{
 		$connection = System::getContainer()->get('database_connection');
-		$userRow = $connection->fetchAssociative('SELECT * FROM tl_user WHERE id = ?', array((int) BackendUser::getInstance()->id));
+		$userRow = $connection->fetchAssociative('SELECT * FROM tl_user WHERE id = ?', array(BackendUser::getInstance()->id));
 
 		if (!$preserveData && $insertOnly)
 		{
